requirements.txt
langchain
langchain-core
python-dotenv
langchain-community
langchain-huggingface
langchain-chroma
langchain-groq
flask
sentence-transformers